refactor(crane-context): split types.ts by domain (plan §R)
Pure type move — no runtime changes. Splits 625-LOC monolith into
9 per-domain files under `workers/crane-context/src/types/`:

- env.ts         (~57 LOC) — Env interface (Worker bindings + secrets)
- sessions.ts    (~130 LOC) — SessionRecord, history blocks, SOD/EOD/Update/Heartbeat request+response types
- handoffs.ts    (~55 LOC) — HandoffRecord, LatestHandoffResponse, HandoffsResponse
- checkpoints.ts (~36 LOC) — CheckpointRecord
- notifications.ts (~57 LOC) — NotificationRecord + re-exported notification enums
- machines.ts    (~52 LOC) — MachineRecord + register/list/heartbeat/ssh-mesh response types
- notes.ts       (~18 LOC) — NoteRecord
- schedule.ts    (~44 LOC) — ScheduleItemRecord, PlannedEventRecord, WorkDayRecord
- internal.ts    (~65 LOC) — IdempotencyKeyRecord, RequestLogRecord, ErrorResponse, ValidationErrorResponse, PaginationCursor, AuthContext, RequestContext
- index.ts       (9 LOC)   — barrel re-export of all domain files

src/types.ts is now a single-line barrel: export * from './types/index.js'

Barrel re-export preserves all existing import paths — no importer changes needed.

Mechanical gate: types.ts drops out of >500 LOC list. No new files >500 LOC.
