+++ workers/crane-context/src/types.ts
@@ -1,625 +1,1 @@
-/**
- * Crane Context Worker - Type Definitions
- *
- * Core types and interfaces for the Crane Context Worker.
- * Matches ADR 025 specification.
- */
-
-import type {
-  Venture,
-  SessionStatus,
-  EndReason,
-  NotificationSeverity,
-  NotificationStatus,
-  NotificationSource,
-  NotificationMatchKeyVersion,
-  NotificationAutoResolveReason,
-} from './constants'
-
-// Re-export for convenience
-export type {
-  Venture,
-  SessionStatus,
-  EndReason,
-  NotificationSeverity,
-  NotificationStatus,
-  NotificationSource,
-  NotificationMatchKeyVersion,
-  NotificationAutoResolveReason,
-}
-
-// ============================================================================
-// Environment Bindings
-// ============================================================================
-
-export interface Env {
-  // Database binding
-  DB: D1Database
-
-  // Configuration (from wrangler.toml vars)
-  CONTEXT_SESSION_STALE_MINUTES: string
-  IDEMPOTENCY_TTL_SECONDS: string
-  HEARTBEAT_INTERVAL_SECONDS: string
-  HEARTBEAT_JITTER_SECONDS: string
-
-  // Feature flags (PR A2 — gates the notification auto-resolver). Set to
-  // 'true' to enable green webhook events to auto-resolve prior failure
-  // notifications. Defaults to disabled until the rollout in PR A4.
-  NOTIFICATIONS_AUTO_RESOLVE_ENABLED?: string
-
-  // Plan v3.1 §D.1: explicit environment marker for /version endpoint
-  // and for smoke-test-e2e.sh mutation-scenario guards. Values:
-  // 'staging' | 'production'. Set via wrangler.toml [vars] and
-  // [env.production.vars]. Missing = 'unknown' in /version output.
-  ENVIRONMENT?: string
-
-  // Memory recall efficacy upgrade (PR 2). Gate on which memories SOS
-  // surfaces as Critical Anti-Patterns. Values:
-  //   'captain_approved' - legacy: require frontmatter.captain_approved
-  //   'injectable'       - require notes.injectable=1 (curator-set)
-  //   'both'             - require BOTH (PR 2 default; bake-in mode)
-  // Bake-in default is 'both' so an unflagged-but-cleanly-curated memory
-  // does NOT inject until the Captain has spot-checked. Flip to
-  // 'injectable' after the 48h observation window. Rollback path is the
-  // same flip back to 'captain_approved' (~30s redeploy).
-  MEMORY_INJECTION_GATE?: string
-
-  // Workers AI binding (added in PR 2 wrangler.toml). Used by the
-  // adversarial-check at write time and the contradiction-axis of the
-  // memory-curator. Optional at type-time so unit tests against the
-  // node:sqlite shim don't have to mock it.
-  AI?: {
-    run(model: string, params: Record<string, unknown>): Promise<unknown>
-  }
-
-  // Secrets (from wrangler secret put)
-  CONTEXT_RELAY_KEY: string
-  CONTEXT_ADMIN_KEY: string
-
-  // Plan v3.1 §B.6 / #454: fine-grained PAT with read-only access to
-  // venturecrane org. Used by the deploy-heartbeats reconciliation
-  // cron (src/deploy-heartbeats-reconcile.ts) to query workflow runs.
-  // If unset, the reconciliation no-ops with a warning log.
-  GH_TOKEN?: string
-
-  // Infisical management token used by /admin/provision-engagement and
-  // /admin/engagement-secrets to create folders and proxy secret reads
-  // for SS engagements. Scope this token to the /ss/clients/* subtree
-  // only; rotate quarterly. See docs/process/new-engagement-setup-checklist.md.
-  // If unset, the provisioning/secrets-proxy endpoints return 503.
-  INFISICAL_MANAGEMENT_TOKEN?: string
-}
-
-// ============================================================================
-// Database Records (as stored in D1)
-// ============================================================================
-
-export interface SessionRecord {
-  // Identity
-  id: string
-
-  // Session context
-  agent: string
-  client: string | null
-  client_version: string | null
-  client_session_id: string | null
-  host: string | null
-  venture: string
-  repo: string
-  track: number | null
-  issue_number: number | null
-  branch: string | null
-  commit_sha: string | null
-
-  // Lifecycle
-  status: SessionStatus
-  created_at: string
-  started_at: string
-  last_heartbeat_at: string
-  ended_at: string | null
-  last_activity_at: string | null
-  end_reason: EndReason | null
-
-  // Grouping (for parallel agent awareness)
-  session_group_id: string | null
-
-  // Schema versioning
-  schema_version: string
-
-  // Attribution & tracing
-  actor_key_id: string
-  creation_correlation_id: string
-
-  // Extensibility
-  meta_json: string | null
-}
-
-export interface HandoffRecord {
-  // Identity
-  id: string
-
-  // Linkage
-  session_id: string
-
-  // Context
-  venture: string
-  repo: string
-  track: number | null
-  issue_number: number | null
-  branch: string | null
-  commit_sha: string | null
-
-  // Handoff metadata
-  from_agent: string
-  to_agent: string | null
-  status_label: string | null
-  summary: string
-
-  // Payload
-  payload_json: string
-  payload_hash: string
-  payload_size_bytes: number
-  schema_version: string
-
-  // Attribution & tracing
-  created_at: string
-  actor_key_id: string
-  creation_correlation_id: string
-}
-
-export interface CheckpointRecord {
-  // Identity
-  id: string
-
-  // Linkage
-  session_id: string
-
-  // Context
-  venture: string
-  repo: string
-  track: number | null
-  issue_number: number | null
-  branch: string | null
-  commit_sha: string | null
-
-  // Checkpoint content
-  summary: string
-  work_completed: string | null // JSON array
-  blockers: string | null // JSON array
-  next_actions: string | null // JSON array
-  notes: string | null
-
-  // Metadata
-  checkpoint_number: number
-  created_at: string
-
-  // Attribution & tracing
-  actor_key_id: string
-  correlation_id: string
-}
-
-export interface IdempotencyKeyRecord {
-  // Composite key
-  endpoint: string
-  key: string
-
-  // Response storage
-  response_status: number
-  response_hash: string
-  response_body: string | null
-  response_size_bytes: number
-  response_truncated: number // 0 = false, 1 = true
-
-  // TTL
-  created_at: string
-  expires_at: string
-
-  // Attribution
-  actor_key_id: string
-  correlation_id: string
-}
-
-export interface RequestLogRecord {
-  id: string
-
-  // Request metadata
-  timestamp: string
-  correlation_id: string
-  endpoint: string
-  method: string
-
-  // Context
-  actor_key_id: string
-  agent: string | null
-  venture: string | null
-  repo: string | null
-  track: number | null
-  issue_number: number | null
-
-  // Response
-  status_code: number
-  duration_ms: number
-  error_message: string | null
-
-  // Idempotency
-  idempotency_key: string | null
-  idempotency_hit: number // 0 = false, 1 = true
-}
-
-// ============================================================================
-// Machine Registry Records
-// ============================================================================
-
-export interface MachineRecord {
-  id: string
-  hostname: string
-  tailscale_ip: string
-  user: string
-  os: string
-  arch: string
-  pubkey: string | null
-  role: string
-  status: string
-  registered_at: string
-  last_seen_at: string
-  meta_json: string | null
-  actor_key_id: string
-}
-
-// ============================================================================
-// Notes Records
-// ============================================================================
-
-export interface NoteRecord {
-  id: string
-  title: string | null
-  content: string
-  tags: string | null // JSON array
-  venture: string | null
-  archived: number // 0 = false, 1 = true
-  created_at: string
-  updated_at: string
-  actor_key_id: string | null
-  meta_json: string | null
-  // Provenance + curator fields (migration 0044, populated by PR 2 curator)
-  authored_by_session_id?: string | null
-  source_hash?: string | null
-  embedding_model?: string | null
-  embedding_version?: string | null
-  embedding_hash?: string | null
-  injectable?: number // 0 = false, 1 = true; default 0
-}
-
-// ============================================================================
-// Schedule Records
-// ============================================================================
-
-export interface ScheduleItemRecord {
-  id: string
-  name: string
-  title: string
-  description: string | null
-
-  cadence_days: number
-  scope: string
-  priority: number
-
-  last_completed_at: string | null
-  last_completed_by: string | null
-  last_result: string | null
-  last_result_summary: string | null
-
-  gcal_event_id: string | null
-
-  enabled: number // 0 = false, 1 = true
-  created_at: string
-  updated_at: string
-}
-
-// ============================================================================
-// Planned Event Records
-// ============================================================================
-
-export interface PlannedEventRecord {
-  id: string
-  event_date: string
-  venture: string
-  gcal_event_id: string | null
-  title: string
-  start_time: string
-  end_time: string
-  type: 'planned' | 'actual' | 'cancelled'
-  sync_status: 'pending' | 'synced' | 'error'
-  created_at: string
-  updated_at: string
-}
-
-export interface SessionHistoryBlock {
-  start: string
-  end: string
-  session_count: number
-  hosts: string[]
-  repos: string[]
-  branches: string[]
-  issues: number[]
-}
-
-export interface SessionHistoryEntry {
-  venture: string
-  work_date: string
-  blocks: SessionHistoryBlock[]
-  total_sessions: number
-}
-
-// ============================================================================
-// Work Day Records
-// ============================================================================
-
-export interface WorkDayRecord {
-  date: string
-  gcal_event_id: string | null
-  started_at: string
-  ended_at: string | null
-  created_at: string
-  updated_at: string
-}
-
-// ============================================================================
-// Notification Records
-// ============================================================================
-
-export interface NotificationRecord {
-  id: string
-  source: NotificationSource
-  event_type: string
-  severity: NotificationSeverity
-  status: NotificationStatus
-  summary: string
-  details_json: string
-  external_id: string | null
-  dedupe_hash: string
-  venture: string | null
-  repo: string | null
-  branch: string | null
-  environment: string | null
-  created_at: string
-  received_at: string
-  updated_at: string
-  actor_key_id: string
-
-  // Match-key fields (added in migration 0023 for the auto-resolver).
-  // Nullable for backward compatibility with rows inserted before the
-  // PR A2 code path was deployed. Legacy rows are backfilled in-migration
-  // from details_json where possible.
-  workflow_id: number | null
-  workflow_name: string | null
-  run_id: number | null
-  head_sha: string | null
-  check_suite_id: number | null
-  check_run_id: number | null
-  app_id: number | null
-  app_name: string | null
-  deployment_id: string | null
-  project_name: string | null
-  target: string | null
-  match_key: string | null
-  match_key_version: NotificationMatchKeyVersion | null
-  run_started_at: string | null
-  auto_resolved_by_id: string | null
-  auto_resolve_reason: NotificationAutoResolveReason | null
-  resolved_at: string | null
-}
-
-// ============================================================================
-// API Request Types
-// ============================================================================
-
-export interface SODRequest {
-  schema_version: string
-  agent: string
-  client?: string
-  client_version?: string
-  host?: string
-  venture: Venture
-  repo: string
-  track?: number
-  issue_number?: number
-  branch?: string
-  commit_sha?: string
-}
-
-export interface EODRequest {
-  schema_version: string
-  session_id: string
-  handoff: {
-    summary: string
-    status_label?: string
-    to_agent?: string
-    work_completed?: string[]
-    blockers?: string[]
-    next_actions?: string[]
-    [key: string]: unknown // Allow additional fields
-  }
-}
-
-export interface UpdateRequest {
-  schema_version: string
-  session_id: string
-  branch?: string
-  commit_sha?: string
-  meta?: Record<string, unknown>
-}
-
-export interface HeartbeatRequest {
-  schema_version: string
-  session_id: string
-}
-
-// ============================================================================
-// API Response Types
-// ============================================================================
-
-export interface SODResponse {
-  session: {
-    id: string
-    status: SessionStatus
-    created_at: string
-    last_heartbeat_at: string
-    schema_version: string
-    venture: string
-    repo: string
-    track?: number
-    issue_number?: number
-  }
-  last_handoff?: {
-    id: string
-    summary: string
-    status_label?: string
-    created_at: string
-    from_agent: string
-  }
-  active_sessions?: Array<{
-    agent: string
-    track?: number
-    issue_number?: number
-    last_heartbeat_at: string
-  }>
-}
-
-export interface EODResponse {
-  session_id: string
-  handoff_id: string
-  ended_at: string
-}
-
-export interface UpdateResponse {
-  session_id: string
-  updated_at: string
-}
-
-export interface HeartbeatResponse {
-  session_id: string
-  last_heartbeat_at: string
-  next_heartbeat_at: string
-  heartbeat_interval_seconds: number
-}
-
-export interface ActiveSessionsResponse {
-  sessions: Array<{
-    id: string
-    agent: string
-    venture: string
-    repo: string
-    track?: number
-    issue_number?: number
-    status: SessionStatus
-    last_heartbeat_at: string
-    created_at: string
-  }>
-  pagination?: {
-    next_cursor?: string
-  }
-}
-
-export interface LatestHandoffResponse {
-  handoff?: {
-    id: string
-    session_id: string
-    from_agent: string
-    to_agent?: string
-    summary: string
-    status_label?: string
-    payload: Record<string, unknown>
-    created_at: string
-  }
-}
-
-export interface HandoffsResponse {
-  handoffs: Array<{
-    id: string
-    session_id: string
-    from_agent: string
-    summary: string
-    created_at: string
-  }>
-  pagination?: {
-    next_cursor?: string
-  }
-}
-
-// ============================================================================
-// Error Response Types
-// ============================================================================
-
-export interface ErrorResponse {
-  error: string
-  details?: unknown
-  correlation_id?: string
-}
-
-export interface ValidationErrorResponse {
-  error: 'validation_failed'
-  details: Array<{
-    field: string
-    message: string
-    params?: Record<string, unknown>
-  }>
-  correlation_id: string
-}
-
-// ============================================================================
-// Machine Registry Request/Response Types
-// ============================================================================
-
-export interface RegisterMachineRequest {
-  hostname: string
-  tailscale_ip: string
-  user: string
-  os: string
-  arch: string
-  pubkey?: string
-  role?: string
-  meta?: Record<string, unknown>
-}
-
-export interface RegisterMachineResponse {
-  machine: MachineRecord
-  created: boolean
-}
-
-export interface ListMachinesResponse {
-  machines: MachineRecord[]
-  count: number
-}
-
-export interface MachineHeartbeatResponse {
-  id: string
-  hostname: string
-  last_seen_at: string
-}
-
-export interface SshMeshConfigResponse {
-  config: string
-  machine_count: number
-  generated_for: string
-}
-
-// ============================================================================
-// Internal Types
-// ============================================================================
-
-export interface PaginationCursor {
-  timestamp: string // ISO 8601
-  id: string // ULID
-}
-
-export interface AuthContext {
-  actorKeyId: string
-  correlationId: string
-}
-
-export interface RequestContext extends AuthContext {
-  startTime: number
-  endpoint: string
-  method: string
-}
+export * from './types/index.js'
